backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
python-multipart
google-generativeai
firebase-admin
langchain
langchain-text-splitters
PyMuPDF
python-docx
python-pptx
Pillow
easyocr
numpy
certifi
dnspython
httpx
beautifulsoup4
